requirements-basic.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
supabase
httpx